+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * make `Fill` discriminated by its type (BREAKING: removes `FillType` enum)
+* fix callback error type
 
 ## v2.0.0
 * Global type definitions removed, the usage of global types is an anti pattern